+++ src/gamestate.h
@@ -7,7 +7,7 @@ public:
   GameState();
 
   virtual void onClick(int x, int y);
-  virtual void update(int dt);
+  virtual void update(float dt);
   virtual void draw();
 };
